Initial commit
For the most part, this should be a fully working utility once compiled.
I eschewed writing my own GitHub API client since the one from Google
offered all the functionality I needed.

In general, the primary detractor for this utility right now from a
user's point of view is some shoddy documentation at the command line.

In terms of the code itself, the code needs more documentation and tests
(more as in there is hardly any right now). There are also a lot of
opportunities for refactoring code to reduce repetition.