Fix CircleCI build configuration for releases
The build configuration for releases is only supposed to be used with
tags beginning with "v". Unfortunately, due to the onerous nature of
CircleCI 2.0's workflow configuration semantics, the the release phase
is running on all commits because it does not filter out branch-level
commits.

I fixed this by specifying that the release phase should ignore _all_
branches. This means that it will only get built when a tag is a pushed.