Prevent word-splitting in Makefile sh expansions
GNU make will use `sh` to execute shell functions, and `sh` will perform
"word splitting" after parameter expansion. What this means is that if
any word breaks are found in the value of an expanded paramters, the
value will be broken up into multiple values based on those breaks.

In practice, this meant that our release notes were expanded as an array
of arguments instead of a single argument.

Word splitting is prevented by placing the parameter to be expanded in
double quotes.